services/image-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
redis==5.0.1
python-multipart==0.0.6
httpx==0.25.2
pillow==10.1.0
aiofiles==23.1.0
prometheus-client==0.19.0

# AI providers
openai==1.3.7

# Image processing and computer vision
opencv-python-headless==4.8.1.78
numpy==1.24.3

# Additional async support
aiohttp==3.9.1